consulting_firm/requirements.txt
python-dotenv
python-docx
graphviz
reportlab
requests
pytest
# Optional: openai, ollama integration can be enabled via MODEL_PROVIDER and environment
# openai
weasyprint>=60.0
tinycss2>=1.2.0
pyphen>=0.14.0
cairocffi>=1.6.0
cairosvg>=2.7.0
markdown
streamlit>=1.37.0
